wrappers, overlord/snapstate/backend: make link-snap clean up on failure. (#3694)
The `link-snap` task involves multiple steps, each of which can fail. While our state machine deals well with undoing the tasks that led up to a failure, each individual task is assumed to be boolean: it either fails, or it succeeds, in its entirety. This means that each handler needs to clean up after itself when it fails. We're not very good at doing this, and in particular `link-snap` can leave behind a lot of artifacts when it fails. This fixes that.

For simplicity I've propagated the requirement for cleaning up behind oneself into the `wrappers` helper lib. This way each layer involved only needs to know about cleaning up the steps visible at that layer.