tests: use dnf --refresh install to avert stale cache
While our package abstraction helper for spread tests does call dnf
makecache evidence in test failures shows that something is not quite
right and we fail to install packages that have since, in the archive,
been upgraded to a new revision and we attempt to install the old
version oblivious to that fact.

Neal Gompa suggested that we use the --refresh option on dnf install to
always ensure that the package cache is updated if it is out of date.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>