wrappers: symlink completion snippets when symlinking binaries
This also modifies complete.sh to support this, as well as some minor
changes to dirs and snap to accommodate.

Symlinking the snippets is itself simple enough that I didn't think it
was worth creating an ad-hoc helper, so it's just a few lines glommed
onto the end of the existing wrapper funcs (and tests).

    To be clear, I did write the helper, but between it being very few
    lines of code, it needing to return enough information to the
    caller to clean up behind it if something else went wrong, and
    especially as the other user of that helper would be the aliases
    and they don't have access to the AppInfo (which would be the
    obvious thing to pass to such helpers), I decided against it.

How aliases are going to play into completion is still an open
question on the implementation side, which I need to look into soon,
but this is not that work.