many tests: move all panicing fake store methods to a common place
Everything that needed a fake store in unit tests was doing its own
thing, stubbing out the unwanted methods with panics. We had four such
things in the code. This implements a single all-panicing store (in
store/storetest), and uses that as the base for the other stores.

The big win of this is that when you add something to
snapstate.StoreService you just have to stub it in one place instead
of N. As long as you remember ... :-)