corecfg: handle unknown keys that are explicitly set
When updateKeyValueStream is processing a stream with a key that is not
known but is explicitly listed in new configuration it would needlessly
indicate that the file needs writing.

The test that was checking this was flaky because the writes happened
so quickly we were never given a chance to observe the failure
(usually). Running the test repeatedly on my SSD makes it fail very
frequently tough.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>