tests: wait until the port is listening after start the fake store
This change is to fix the tests refresh-all and refresh-all-undo by
waiting util the fake store is ready to be invoqued.

Based on different tests done on PR 3768, 3770 and 3771 it is clear the
problem is because the store is not ready when snap refresh is executed
and it is causing the error seeying in the autopkgtests.