interfaces/backends: enable apparmor backend on parital feature set
AppArmor may be fully or only partially supported on a given system. In
the past when partially supported apparmor was detected then the backend
would not be loaded, leading to very confusing and error-prone behavior
for people switching between mainline (unpatched) and ubuntu (patched)
kernels. Because the backend was not loaded then now-stale profiles were
*not* updated or removed and things behaved oddly, where incorrect
profiles were actually loaded into the kernel (apparmor tries to degrade
gracefully but the profiles were not permissive enough).

With this patch the apparmor backend is loaded as long as the kernel
module is loaded and enabled. This will be coupled with a patch to the
backend, which will generate more permissive profiles for
anything-less-than-full feature set. This patch will allow snapd to
enable apparmor on Debian, openSUSE as well as Ubuntu running pre-4.15
mainline kernels.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>