packaging: unify PATH and XDG_DATA_DIRS handling cross-distro
All distributions need to add the snappy bin directory (/snap/bin or
/var/lib/snapd/snap/bin, depending) to the user's PATH, and to add
/var/lib/snapd/desktop to XDG_DATA_DIRS (which might be unset, meaning
it needs adding to the defaults).

Up to now all distros were doing this separately, so we had a bunch of
duplication. Also, as things move to Wayland, on debian-based distros
XDG_DATA_DIRS was only being set in X sessions which mean that
switching to wayland lost you your snappy desktop files.

This tries to fix all that, hopefully in a sane way.