spread: don't set HTTPS?_PROXY for linode (#3820)
* spread: don't set HTTPS?_PROXY for linode

I'm using a proxy at home, this includes me having SPREAD_HTTPS?_PROXY
environment variables set. This works perfectly for qemu or other local
targets but fails miserably for linode. Since using a proxy in linode is
unlikely to be useful this patch disables those two variables for that
backend.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* spread: document Linode proxy unsetting

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>