cmd/snap-confine: use child profile for snap-update-ns
This patch allows snap-confine to run snap-update-ns with a very
specific child profile, that currently mirrors (actually moves) the same
mount rules that used to be allowed for snap-confine. At the same time
snap-confine can no longer do that.

The rules are incomplete and I will probably revise this patch.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>