Run the unittests everywhere
We had an issue during the 2.27 cycle where the unittests would
fail during build on 14.04 systems. We disable unit tests during
the package build in the spread tests and our tests/unit/go test
is currently only run on ubuntu-16.04-64 systems.

This PR enables the unit tests everywhere to avoid that we get
caught an issue like the above again.