many: configure store from state, reconfigure store at runtime (#3780)
Configures the store's API from the state - recorded as a new "store": {"base-url": "<url>"} section in the state's data. Also adds a func to reconfigure the store to use a new base API URL at runtime.

The cached Store was previously managed by snapstate but that's moved to a new storestate package, alongside other store-related state helpers.

Note: caching the auth context in the state is a bit of a hack, but is the only sane way I could find (i.e. without affecting lots of code & tests) to make it available so the store's API can be changed by a call to a storestate func instead of going through the overlord.