tests: add new test that checks that the compat snapd-xdg-open works
We are not currently testing that snapd-xdg-open works correctly.
This has lead to a regression with the core snap when a typo broke
the compat sending of the dbus message for the old snapd-xdg-open.

Part of the problem is that we cannot do stable updates with an
unchanged core, we always need to pull in whatever is in the PPA.
This test will ensure the launcher handling does not regress again.