interfaces/opengl: don't udev tag nvidia devices and use snap-confine instead
/dev/nvidia* devices are created via the proprietary nvidia module. Proprietary
nvidia modules are not allowed to use sysfs, therefore they cannot be udev
tagged and ultimately won't be added to the device cgroup when the opengl
interface is connected. For now, adjust snap-confine to add /dev/nvidia* to the
device cgroup if they exist, and let AppArmor handle the mediation.