overlord/snapstate: prefer a smaller corner case for doing the wrong thing
Currently, if a snap is installed with any flags, refreshed
(preserving those flags), and reverted, the flags are lost.

This means that a classic snap suddenly finds itself confined in
strict mode (ditto for devmode; ditto again for jailmode). The current
reply to this was "revert needs explicit flags", which is surprising
to most users. The reasoning being that a snap can be classic at
revision N, and strict at revision M, and as flags are per snap and
not per revision we can't make a sensible decision.

> The  right fix is to make flags per revision, obviously (in retrospect)

While we find time to implement the right fix, this PR changes the
behaviour so that explicit flags are needed to _change_ the flags that
the snap is installed with; if nothing is specified, flags are
preserved.

This will break the aforementioned flip-flopping snaps, but preserve
the apparently more common case. So the breakage is moved from a
bigger corner case to a smaller one, which is a win.

This builds on #3940.