interfaces: allow backends to initialize
This patch adds a dummy Initialize method to all security backends.
In the future certain backends will perform one-off initialization to
ensure they can be used correctly. Currently that is done in a hackish
way by abusing Setup. This is slightly incorrect as it relies on us
actually doing something that causes snap-specific security setup to be
required. We want to decouple setting up snap security and setting up
security systems so that each responsibility is clear.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>