snap: introduce structured epochs.
Structured epochs are backwards-compatible with the epochs we had, but
allow to express scenarios that the old ones did not.

From the comment on the Epoch type,

>     epoch: N
>
> means a snap can read/write exactly the Nth epoch's data, and
>
>     epoch: N*
>
> means a snap can additionally read (N-1)th epoch's data, which means it's a
> snap that can migrate epochs (so a user on epoch 0 can get offered a refresh
> to a snap on epoch 1*).

this is as things were. And now,

> If the above is not enough, a developer can explicitly describe what epochs a
> snap can read and write:
>
>     epoch:
>       read: [1, 2, 3]
>       write: [1, 3]
>
> the read attribute defaults to the value of the write attribute, and the
> write attribute defaults to the last item in the read attribute. If both are
> unset, it's the same as not specifying an epoch at all (i.e. epoch: 0). The
> lists must be in ascending order, and there must be a non-empty intersection
> between them.

The current commit tries to serialise epochs in the older format
first, and only uses the new format if it can't be done. This keeps
things nice and simple for the common case, and has the additional
benefit of keeping the store happy until things are implemented there.