osutil: add code for FindUid/FindGid from snap-seccomp
This patch adds an easily accessible pair of functions for performing
user and group lookups. Group lookups are not in golang 1.6 so we cherry
picked part of the future implementation but now we need to use them
from more than once place, hence the new entry in the osutil package.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>