interfaces: Enhance full-confinement support for biarch distributions
This change adds support for those using the "Arch Linux" (biarch) style
symlink farms for NVIDIA driver support when using full apparmor confinement.

Additionally this change adds support for the linux-driver-management method for
managing the host-side drivers, whereby the `glx-provider/*/` directories
are populated with the real library assets, with symlinks in the main
library directories being pointed to these locations.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>