tests: expand the freezer cgroup test
This patch ensures that multiple processes can inhabit a freezer cgroup
as well as that a process can join an existing cgroup. The test code is
also more correct as "sleep 1h &" was changed to "exec sleep 1h &" which
ensures that only one background process is active. The terminated
processes are now also correctly waited for.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>