Fix revert test when a new core image is on current channel
This fix is to deal with the scenario when a new core is pushed during
the tests execution, and if "snap refresh" is executed, the new core is
installed and it reboots the machine making the whole suite fail.
This is reproduced when we run on edge channel with external backend.