overlord/snapstate: support completion for command aliases
This makes the code that manages command aliases aware of completion
support needs. That is: on alias CRUD, also CRUD the appropriate
symlinks for the completion snippets of the aliases.

For example, if you have a snap `foo` that has an app `bar`, and
`foo.bar` has a `completer` keyword in the yaml, then typing
`foo.bar <tab>` in bash will offer you completion via that snippet. If
you then `snap alias foo.bar bar` you'd expect to see the same
completion suggestions. This PR hooks things up so that you would, in
fact, get them -- but the completion script needs to support this!

That is, if your completer is something like

```sh
_complete_bar() {
   COMPREPLY=(hello world)
}
complete -F _complete_bar foo.bar
```

then things will work for `foo.bar` but *not* for the `bar` alias. For
any alias that your snap uses, the completer snippet itself should
list it:

```sh
_complete_bar() {
   COMPREPLY=(hello world)
}
complete -F _complete_bar foo.bar bar
```

Or, if it wants to support completing with any command,
declare the function as the default completer with `-D`:

```sh
_complete_bar() {
   COMPREPLY=(hello world)
}
complete -D _complete_bar
```

(as this runs confined and in its own process it won't take over the
default completer).

HTH, HAND.