cmd/snap-update-ns: reject mounting affecting symlinks
Unfortunately mount cannot be made not to follow symlinks so we take
extra measures to ensure that we don't even use mount with something
that is not a directory. Apart from freezing all the processes we now
lstat the target (for regular mounts) or both source and target (for
bind mounts), to ensure that they are directories and not anything else.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>