Fix econnreset test for devices when retry is needed to download
In the tests for beta validation this test is often failing when the
retry is triggered. To deal with that, I am adding extra time to wait
until the partial file is created and the download is started.

Also, the snap process which is downloading the huge file is killed to
avoid the script fails with a timeout dalaying the whole suite
execution.

For an example see:
https://paste.ubuntu.com/25717714 line 1191
https://paste.ubuntu.com/25719948/ line 388