cmd/snap-update-ns: chown created directory segments
This patch changes SecureMkdirAll to chown each created path segment.
This is necessary as snap-update-ns may run with uid 0 and gid of a
regular user and in result leave behind directories with group ownership
of the calling user.

Because chown is now integrated into SecureMkdirAll the code and tests
were simplified to remove explicit handling of chown mocking.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>