store: simply cache and remove digest handling
The download cache was previously "smart" and would do content
addressing internally. However it turns out that this is not
really needed. The store knowns the hashes already and the cache
itself does not need to have any knowledge about this. So the
cache is now simplified and knows only about "cacheKeys" which
can be arbitrary from the caches POV.