tests: create /etc/{group,gshadow,passwd,shadow} for pristine core
On a unmodified core snap under testing spread wants to login as
root in the linode backend. However root is of course disabled
on a pristine core snap. In the past we modified the core snap
itself to allow root login. However this will not work when trying
to test going from e.g. "stable" core to "edge" core as we need
to test the snaps from the store. To make this work, create
/var/lib/extrausers/{group,gshadow,passwd,shadow} based on the
core snap but with the injected "root,ubuntu,test" users from
the stock system. Then bind mount those on boot to /etc/{group,
gshadow,passwd,shadow} to allow spread/linode to work.