interfaces/builtin: add support for content "source" section
This patch alters the content interface to support the new syntax using
"source" section on content slots. This allows a content slot to share a
number of entries (both read and write at the same time).

When a "source" section is in use, the target directory on the plug
becomes a spool-like directory where content connections from various
connected slots is aggregated.

This change itself is simple but will only work after the automatic
directory creation and hole poking patches are merged. This patch also
needs one more feature to allow the mount backend to de-duplicate mount
entries when two slots connected to one plug wish to use the same
directory name. This will be done in a separate patch.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>