systemd: run all mount units before snapd.service to avoid race
There is a race on startup in Ubuntu 14.04. The mount units for
the installed snaps and the systemd unit are started at similar
times and sometimes the mount units are not ready when snapd
starts. This results in confusing errors like snaps loose all
their interfaces because the information about the interfaces
is stored in the on-disk snap.yaml only (reported in
LP: #1721518).

This PR fixes this issue by adding a "Before=snapd.service"
relation to all the snap mount units.