wrappers: do not error when a desktop file has invalid Exec= lines
So far we ignored errors in the Exec= lines for snaps. This changed
recently however it broke too many things in the wild so we needed
to revert that. Instead of failing hard we now try to fix the
Exec= line by guessing what snap app to use (based on the desktop
file) as suggested by Gustavo. In addition we log the error into
syslog.