add spread test for connecting all interfaces (excepting gadget slots)
Test each interface for app and os snaps and verify a simple can be run when
connected. This provides high-level testing for things like AppArmor policy
syntax errors, seccomp policy parsing, udev querying bugs, etc.

This provides two provider snaps: test-policy-app-provider-core and
test-policy-app-provider-classic and chooses which to install based on if the
system is a core system or not.

It also provides the test-policy-app-consumer snap that is used to connect to
the provider snap for app-provided slots and to the core snap for implicit
slots.

The provider and consumer snaps in tests/lib/snaps should be updated when new
interfaces are added.