ifacestate: make interfaces.Repository available via state cache
This PR makes the interfaces.Repository available for other managers
via the state cache. This is similar to how the assertdb is made
available. This allows e.g. snapstate to get the interfaces
repository and answer questions like "is this plug connected".

The code to get/set the cached repository is in it's own sub-package
to avoid circular imports between snapstate and ifacestate.