tests: fix security-device-cgroup* tests on devices with framebuffer
The tests that check if the device cgroup works for framebuffer
devices is broken currently on devices that have an active
framebuffer. We did not notice so far because all virtual machines
we use to test do not have a framebuffer device.

However devices like the pi2/pi3 have framebuffers. There it is
not enough to create /dev/fb255 because the cgroup for framebuffer
device is active (it contains fb0). On machines that do not have
any framebuffer device the device cgroup is not active and we only
test that the apparmor based access is working.

The fix is to always use "fb0" in the tests and only create it
if it is missing. On systems were it is missing we tests the
apparmor based permissions as before. But on systems with a
framebuffer we now test also the right device in the cgroup
so that the test can actually work, i.e. snap-confine will
add the right (fb0) udev device to the devices cgroup. The
artificially created fb255 that was used before was not known
to udev and therefore not added by snap-confine to the devices
cgroup.