cmd/snap-confine: add slave PTYs and let devpts newinstance perform mediation
Unconditionally add /dev/pts slave devices (we already add /dev/ptmx) to the
device cgroup so that the slaves are available when interfaces that trigger
the device cgroup are in effect. We use a glob for minor (eg, 'c 136:* rwm')
since we don't know what future slaves will be created (this is ok because we
use devpts newinstance).

Also cleanup the devpts spread test. Before it was using expect to create a
python script to run and test the output. Instead, create a test-devpts snap
with two python scripts and run those directly. The snap plugs an interface
that triggers the device cgroup (physical-memory-observe since /dev/mem should
be available everywhere) and runs the test commands with the interface
connected and disconnected to verify we can use devpts if the device cgroup is
in effect or not.

Reference:
https://forum.snapcraft.io/t/how-to-open-pseudoterminals-from-dev-pts/2637/