interfaces/raw-usb: match on SUBSYSTEM, not SUBSYSTEMS
The raw-usb interface is intended to give access to all USB devices in
/dev/bus/usb. Part of that included udev tagging to add these devices to the
snap-specific device cgroup, and it did so using 'SUBSYSTEMS=="usb", ...'.
While this worked on newer systems, on at least Ubuntu Core 16 this caused a
udev_enumerate error. SUBSYSTEMS isn't correct anyway because it means to
include parent devices which in our case may or may not be USB and therefore
the security policy wouldn't allow those devices anyway. Update this to use
'SUBSYSTEM=="usb", ...' instead, which limits the tagging to just the USB
devices (not their parents) and works on Ubuntu Core 16.