cmd/snap-confine: Implement full 32-bit NVIDIA driver support
This refactors mount-support-nvidia somewhat to allow us to expose the
32-bit NVIDIA drivers (from `/usr/lib32`) to the guest snap on both multiarch
and biarch systems, in their `/var/lib/snapd/lib/gl32` directory.

The `SNAP_LIBRARY_PATH` is modified to include the new directory, so that
32-bit processes in normal snapcraft packages using the helpers do not need
to do anything to benefit from them.

This particularly benefits the LSI project by making the 32-bit drivers
available to Steam.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>