osutil/user: replace our uses of os/user and filepath.Glob("/home/*")
Previously we were using a mixture of os/user and filepath globs to
get at users in the system. This replaces those uses with an ad-hoc
package, which is initially a very minimal backend that just knows
about files, but is careful to have an interface that should make
replacing it with a bigger, or more modular backend possible down the
road with no impact on the rest of the codebase.

Tangentially to this but also impacting it is golang#22739 which means
os.Getuid() and co use the wrong integer type and are thus wrong for
high ids (especially noticeable on 32-bit arches). This change also
addresses this (the only remaining uses are for comparing with 0,
which is OK), but it still impacts godbus in our dependencies --
anything that uses godbus on a system with high uids is broken today.

There is still some room for things to break: osutil still includes a
cgo-based implementation of group lookup, which should probably be
replaced by something analogous to the current user. AppArmor doesn't
afaik really know about user homes outside of /home/ (except /root?),
and this code will produce such users (this was broken before, in the
opposite direction: users with homes outside of /home/ were partially
invisible; in both cases snapd is unusabe without tweaks to the local
system's rules).