tests/lib/pkgdb, spread: use pacman -S[y]u when installig packages
This is a workaround for Arch's pacman not being smart enough. The way pacman
works is that when a dependency for a package being installed is already
present, it will not be upgraded. Thus it is important that package installation
is done by calling `pacman -Syu ..` which has a side effect of upgrading all
packages in the system.

The patch introduces --sysupgrade option where possible

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>