spread: source prepare-restore.sh in prepare[-each] & restore[-each]
Parts that were previously inside spread.yaml prepare/restore steps are now
moved to prepare-restore.sh script. Since this the script is executed in a
separate shell, the utility functions REBOOT and MATCH are not available. Both
actions may be needed when preparing the test host. Source the script so that we
can regain access to these functions.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>