Fix for issues when snad service is canceled and cleanup for snaps on trusty
Three issues are fixed/improved with change.

The first is how the snapd service is stopped. To avoid the error "Job
canceled" when the snapd service is stopped but it is in "activating"
state. To deal with that, the solution is to wait until the state change
to active, to stop it.

The second is how the services are stopped in the postrm scritps. The
idea of this change is to avoid call "systemctl stop service" many times
when the services take some time to be stopped. This was causing some
services like the lxd or neutron are stopped many times.

The third is that the postrm script will fail in case the final cleanup
is not completed properly.

The last once is a change in the install-snaps test to remove the snap
as part of the test and check the snap is properly removed.